+++ packages/moonstone/Panels/Panel.js
@@ -131,7 +131,6 @@ const PanelBase = kind({
 				}
 
 				Spotlight.set(spotlightId, config);
-				Spotlight.setActiveContainer(spotlightId);
 				Spotlight.focus(spotlightId);
 			}
 		}